Update README.md
This repository contains my Online Store DataBase project. My aim is to show my ability to do certain key clauses.

About:
This is an online company that primarily sells clothes and accessories. The small business owner is asking questions to improve business processes. The dataset is completely made up therefore a link will not be provided.

Summary: The owner wants to know what are the total sales to see if they are making a profit. Customers begin are beginning to ask questions about their orders and the owner needs to figure out how many orders are processing. They realize that the amount of outstanding orders is bad for their store and they need to improve this process.